+++ source/extensions/common/dynamic_forward_proxy/dns_cache_impl.cc
@@ -2,9 +2,9 @@
 
 #include "envoy/extensions/common/dynamic_forward_proxy/v3/dns_cache.pb.h"
 
+#include "common/config/utility.h"
 #include "common/http/utility.h"
 #include "common/network/utility.h"
-#include "common/config/utility.h"
 
 // TODO(mattklein123): Move DNS family helpers to a smaller include.
 #include "common/upstream/upstream_impl.h"
@@ -15,8 +15,8 @@ namespace Common {
 namespace DynamicForwardProxy {
 
 DnsCacheImpl::DnsCacheImpl(
-    Event::Dispatcher& main_thread_dispatcher, ThreadLocal::SlotAllocator& tls, Runtime::RandomGenerator& random,
-    Stats::Scope& root_scope,
+    Event::Dispatcher& main_thread_dispatcher, ThreadLocal::SlotAllocator& tls,
+    Runtime::RandomGenerator& random, Stats::Scope& root_scope,
     const envoy::extensions::common::dynamic_forward_proxy::v3::DnsCacheConfig& config)
     : main_thread_dispatcher_(main_thread_dispatcher),
       dns_lookup_family_(Upstream::getDnsLookupFamilyFromEnum(config.dns_lookup_family())),
@@ -24,9 +24,12 @@ DnsCacheImpl::DnsCacheImpl(
       scope_(root_scope.createScope(fmt::format("dns_cache.{}.", config.name()))),
       stats_{ALL_DNS_CACHE_STATS(POOL_COUNTER(*scope_), POOL_GAUGE(*scope_))},
       refresh_interval_(PROTOBUF_GET_MS_OR_DEFAULT(config, dns_refresh_rate, 60000)),
+      failure_backoff_strategy_(
+          Config::Utility::prepareDnsRefreshStrategy<
+              envoy::extensions::common::dynamic_forward_proxy::v3::DnsCacheConfig>(
+              config, refresh_interval_.count(), random)),
       host_ttl_(PROTOBUF_GET_MS_OR_DEFAULT(config, host_ttl, 300000)),
-      max_hosts_(PROTOBUF_GET_WRAPPED_OR_DEFAULT(config, max_hosts, 1024)), failure_backoff_strategy_(Config::Utility::prepareDnsRefreshStrategy<envoy::extensions::common::dynamic_forward_proxy::v3::DnsCacheConfig>(
-          config, refresh_interval_.count(), random)) {
+      max_hosts_(PROTOBUF_GET_WRAPPED_OR_DEFAULT(config, max_hosts, 1024)) {
   tls_slot_->set([](Event::Dispatcher&) { return std::make_shared<ThreadLocalHostInfo>(); });
   updateTlsHostsMap();
 }
@@ -201,9 +204,10 @@ void DnsCacheImpl::finishResolve(const std::string& host,
   // is populated dynamically.
   if (status == Network::DnsResolver::ResolutionStatus::Success) {
     failure_backoff_strategy_->reset();
-  primary_host_info.refresh_timer_->enableTimer(refresh_interval_);
+    primary_host_info.refresh_timer_->enableTimer(refresh_interval_);
   } else {
-    primary_host_info.refresh_timer_->enableTimer(std::chrono::milliseconds(failure_backoff_strategy_->nextBackOffMs()));
+    primary_host_info.refresh_timer_->enableTimer(
+        std::chrono::milliseconds(failure_backoff_strategy_->nextBackOffMs()));
   }
 }